Move TMDB infra package to persistence
Refactor TMDB code: moved classes from org.tvl.tvlooker.infrastructure.tmdb to org.tvl.tvlooker.persistence.tmdb. Renamed/moved TmdbClient, DTOs, and mapper packages and updated all affected imports in service and mapper classes. Also updated the design docs to reflect the new package paths and applied small formatting fixes (Java record braces and ParameterizedTypeReference anonymous class formatting).
