+++ src/main/java/org/tvl/tvlooker/persistence/tmdb/mapper/TmdbPersonMapper.java
@@ -1,8 +1,8 @@
-package org.tvl.tvlooker.infrastructure.tmdb.mapper;
+package org.tvl.tvlooker.persistence.tmdb.mapper;
 
 import org.tvl.tvlooker.domain.model.entity.Actor;
 import org.tvl.tvlooker.domain.model.entity.Director;
-import org.tvl.tvlooker.infrastructure.tmdb.dto.TmdbCreditsDto;
+import org.tvl.tvlooker.persistence.tmdb.dto.TmdbCreditsDto;
 import org.tvl.tvlooker.persistence.repository.ActorRepository;
 import org.tvl.tvlooker.persistence.repository.DirectorRepository;
