+++ src/main/java/org/tvl/tvlooker/service/tmdb/TmdbDataSynchronizer.java
@@ -12,20 +12,21 @@ import org.tvl.tvlooker.domain.model.entity.Director;
 import org.tvl.tvlooker.domain.model.entity.Genre;
 import org.tvl.tvlooker.domain.model.entity.Item;
 import org.tvl.tvlooker.domain.model.enums.TmdbType;
-import org.tvl.tvlooker.infrastructure.tmdb.TmdbClient;
-import org.tvl.tvlooker.infrastructure.tmdb.dto.TmdbChangesDto;
-import org.tvl.tvlooker.infrastructure.tmdb.dto.TmdbCreditsDto;
-import org.tvl.tvlooker.infrastructure.tmdb.dto.TmdbGenreDto;
-import org.tvl.tvlooker.infrastructure.tmdb.dto.TmdbMovieDto;
-import org.tvl.tvlooker.infrastructure.tmdb.dto.TmdbPagedResponseDto;
-import org.tvl.tvlooker.infrastructure.tmdb.dto.TmdbTvShowDto;
-import org.tvl.tvlooker.infrastructure.tmdb.mapper.TmdbGenreMapper;
-import org.tvl.tvlooker.infrastructure.tmdb.mapper.TmdbItemMapper;
-import org.tvl.tvlooker.infrastructure.tmdb.mapper.TmdbPersonMapper;
+
 import org.tvl.tvlooker.persistence.repository.ActorRepository;
 import org.tvl.tvlooker.persistence.repository.DirectorRepository;
 import org.tvl.tvlooker.persistence.repository.GenreRepository;
 import org.tvl.tvlooker.persistence.repository.ItemRepository;
+import org.tvl.tvlooker.persistence.tmdb.TmdbClient;
+import org.tvl.tvlooker.persistence.tmdb.dto.TmdbChangesDto;
+import org.tvl.tvlooker.persistence.tmdb.dto.TmdbCreditsDto;
+import org.tvl.tvlooker.persistence.tmdb.dto.TmdbGenreDto;
+import org.tvl.tvlooker.persistence.tmdb.dto.TmdbMovieDto;
+import org.tvl.tvlooker.persistence.tmdb.dto.TmdbPagedResponseDto;
+import org.tvl.tvlooker.persistence.tmdb.dto.TmdbTvShowDto;
+import org.tvl.tvlooker.persistence.tmdb.mapper.TmdbGenreMapper;
+import org.tvl.tvlooker.persistence.tmdb.mapper.TmdbItemMapper;
+import org.tvl.tvlooker.persistence.tmdb.mapper.TmdbPersonMapper;
 
 import java.time.LocalDate;
 import java.util.HashSet;
